modify: 백엔드 취급의 퍼센트 타입 모두 Double로 변경

+++ backend/src/main/java/com/duckstar/domain/mapping/WeekAnime.java
@@ -45,7 +45,7 @@ public class WeekAnime extends BaseEntity {
     // 기록용
     private Double votePercent;
 
-    private Integer malePercent;
+    private Double malePercent;
 
     private Integer peakRank;
