+++ backend/src/main/java/com/duckstar/web/dto/AnimeResponseDto.java
@@ -45,8 +45,8 @@ public class AnimeResponseDto {
 
         Integer weeksOnTop10;
 
-        Integer malePercent;
+        Double malePercent;
 
-        Integer femalePercent;
+        Double femalePercent;
     }
 }
